requirements.txt
gradio==4.44.0
opencv-python==4.9.0.80
numpy==1.26.4
scikit-image==0.24.0
Pillow==10.4.0
scikit-learn==1.5.1
